Moving stuff around
Note that planout.rb should be split into multiple files and tested
more thoroughly against Python version
